requirements.txt
torch
torchvision
scikit-learn
numpy
matplotlib
scipy
Cython
pyyaml